split out functions from text-field
