+++ libs/@hashintel/design-system/src/components.ts
@@ -23,6 +23,7 @@ export * from "./icon-input-pipe";
 export * from "./icon-link";
 export * from "./icon-rainbow-hash";
 export * from "./image-with-checked-background";
+export * from "./input-props";
 export * from "./loading-spinner";
 export * from "./menu-checkbox-item";
 export * from "./menu-item";
