+++ libs/@hashintel/design-system/src/text-field.tsx
@@ -1,27 +1,14 @@
-import { faCheckCircle } from "@fortawesome/free-regular-svg-icons";
-import { faCircleExclamation } from "@fortawesome/free-solid-svg-icons";
 import {
   Box,
   Collapse,
   formHelperTextClasses,
-  InputAdornment,
-  inputClasses,
-  InputProps,
-  outlinedInputClasses,
   TextField as MuiTextField,
-  TextFieldProps as MuiTextFieldProps,
   Typography,
   useTheme,
 } from "@mui/material";
 import { forwardRef, FunctionComponent, useState } from "react";
 
-import { FontAwesomeIcon } from "./fontawesome-icon";
-
-type TextFieldProps = {
-  success?: boolean;
-  showLabelCornerHint?: boolean;
-  autoResize?: boolean;
-} & MuiTextFieldProps;
+import { getInputProps, inputLabelProps, TextFieldProps } from "./input-props";
 
 /**
  * 'Freezes' a value when it's falsy, meaning the value will never update to
@@ -34,68 +21,6 @@ const useFrozenValue = <T extends any>(value: T): T => {
     setFrozenValue(value);
   }
   return frozenValue;
-};
-
-export const getInputProps = ({
-  success,
-  variant,
-  error,
-  multiline,
-  autoResize,
-  ...otherProps
-}: InputProps &
-  Pick<
-    TextFieldProps,
-    "success" | "error" | "multiline" | "autoResize" | "variant"
-  > = {}): InputProps => {
-  const { sx: InputPropsSx = [], ...otherInputProps } = otherProps;
-
-  const renderEndAdornment = () => {
-    if (error || success) {
-      return (
-        <InputAdornment position="end">
-          <FontAwesomeIcon
-            icon={success ? faCheckCircle : faCircleExclamation}
-            sx={({ palette }) => ({
-              color: success ? palette.blue[70] : palette.red[60],
-            })}
-          />
-        </InputAdornment>
-      );
-    }
-    return null;
-  };
-
-  return {
-    sx: [
-      ({ palette }) => ({
-        [`& .${outlinedInputClasses.notchedOutline}, &:hover .${outlinedInputClasses.notchedOutline}, &.${inputClasses.disabled} .${outlinedInputClasses.notchedOutline}`]:
-          {
-            borderColor: error ? palette.red[40] : palette.gray[30],
-          },
-        [`&.${inputClasses.disabled}`]: { backgroundColor: palette.gray[20] },
-        ...(multiline &&
-          autoResize && {
-            [`& .${outlinedInputClasses.input}`]: {
-              resize: "auto",
-            },
-          }),
-      }),
-      ...(Array.isArray(InputPropsSx) ? InputPropsSx : [InputPropsSx]),
-    ],
-    /** `notched` is only expected for `outlined` variant, passing it for other variants gives a console warning
-     * @see https://github.com/mui/material-ui/issues/32550 for context
-     */
-    ...(variant === "outlined" ? { notched: false } : {}),
-    ...otherInputProps,
-    endAdornment:
-      error || success ? renderEndAdornment() : otherProps.endAdornment,
-  };
-};
-
-export const inputLabelProps = {
-  disableAnimation: true,
-  shrink: true,
 };
 
 export const TextField: FunctionComponent<TextFieldProps> = forwardRef(
